backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
langgraph==0.2.55
langchain==0.3.9
langchain-aws==0.2.9
langfuse==2.57.3
yfinance==0.2.50
python-jose[cryptography]==3.3.0
boto3==1.35.81
hypothesis==6.122.3
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0
